+++ src/main/kotlin/it/pagopa/wallet/repositories/NpgSessionsTemplateWrapper.kt
@@ -1,16 +1,16 @@
 package it.pagopa.wallet.repositories
 
+import org.springframework.data.redis.core.ReactiveRedisTemplate
 import java.time.Duration
-import org.springframework.data.redis.core.RedisTemplate
 
 class NpgSessionsTemplateWrapper
 /**
  * Primary constructor
  *
- * @param redisTemplate inner redis template
+ * @param reactiveRedisTemplate inner reactive redis template
  * @param ttl time to live for keys
  */
-(redisTemplate: RedisTemplate<String, NpgSession>, ttl: Duration) :
-    RedisTemplateWrapper<NpgSession>(redisTemplate = redisTemplate, "keys", ttl) {
+(reactiveRedisTemplate: ReactiveRedisTemplate<String, NpgSession>, ttl: Duration) :
+    ReactiveRedisTemplateWrapper<NpgSession>(reactiveRedisTemplate = reactiveRedisTemplate, "keys", ttl) {
     override fun getKeyFromEntity(value: NpgSession): String = value.orderId
 }
