+++ src/main/kotlin/it/pagopa/wallet/repositories/UniqueIdTemplateWrapper.kt
@@ -1,17 +1,17 @@
 package it.pagopa.wallet.repositories
 
+import org.springframework.data.redis.core.ReactiveRedisTemplate
 import java.time.Duration
-import org.springframework.data.redis.core.RedisTemplate
 
 class UniqueIdTemplateWrapper
 
 /**
  * Primary constructor
  *
- * @param redisTemplate inner redis template
+ * @param reactiveRedisTemplate inner reactive redis template
  * @param ttl time to live for keys
  */
-(redisTemplate: RedisTemplate<String, UniqueIdDocument>, ttl: Duration) :
-    RedisTemplateWrapper<UniqueIdDocument>(redisTemplate = redisTemplate, "uniqueId", ttl) {
+(reactiveRedisTemplate: ReactiveRedisTemplate<String, UniqueIdDocument>, ttl: Duration) :
+    ReactiveRedisTemplateWrapper<UniqueIdDocument>(reactiveRedisTemplate = reactiveRedisTemplate, "uniqueId", ttl) {
     override fun getKeyFromEntity(value: UniqueIdDocument): String = value.id
 }
